agentic_codebase_genius/BE/requirements.txt
jaclang>=0.8
fastapi>=0.110
uvicorn>=0.30
pydantic>=2.6
python-dotenv>=1.0
GitPython>=3.1
tree_sitter>=0.21.3
graphviz>=0.20.3
networkx>=3.2
requests>=2.32
jac-cloud>=0.0.4
